# PERF: groupby.transform broadcast perf