--- a/asv_bench/benchmarks/groupby.py
+++ b/asv_bench/benchmarks/groupby.py
@@ -741,6 +741,12 @@ class Transform:
             index=np.random.choice(range(10), n),
         )
 
+        n = 1_000_000
+        self.df_tall = DataFrame(
+            np.random.randn(n, 3),
+            index=np.random.randint(0, 5, n),
+        )
+
         n = 20000
         self.df1 = DataFrame(
             np.random.randint(1, n, (n, 3)), columns=["jim", "joe", "jolie"]
@@ -759,6 +765,9 @@ class Transform:
 
     def time_transform_ufunc_max(self):
         self.df.groupby(level="lev1").transform(np.max)
+
+    def time_transform_lambda_max_tall(self):
+        self.df_tall.groupby(level=0).transform(lambda x: np.max(x, axis=0))
 
     def time_transform_lambda_max_wide(self):
         self.df_wide.groupby(level=0).transform(lambda x: np.max(x, axis=0))