@lukemanley 2022-01-30
@phofl - there are not many groupby transform asvs that hit this code path. I just added one for this case.