@jreback 2022-01-30
> We like to target ASVs to be around 80ms, is that right? Just want to make sure, looks appropriate as-is.

yeah generally < 100ms is good, but we have a lot that are not so can cleanup separately.